+++ pueue/src/client/commands/clean.rs
@@ -11,11 +11,13 @@ pub async fn clean(
     style: &OutputStyle,
     group: Option<String>,
     successful_only: bool,
+    older_than: Option<u64>,
 ) -> Result<()> {
     client
         .send_request(CleanRequest {
             successful_only,
             group,
+            older_than,
         })
         .await?;